+++ CHANGELOG.md
@@ -113,6 +113,9 @@ Adding a new version? You'll need three changes:
  - Added support for drain support functionality.
    It can be enabled by setting the flag `--enable-drain-support` to `true`.
    [#7466](https://github.com/Kong/kubernetes-ingress-controller/pull/7466)
+ - Store the license fetched from Konnect to the `Secret` named `konnect-license-<cpID>`
+   in the same namespace where KIC runs. The `cpID` is the ID of the Konnect
+   control plane. This requires the `Secret` to be created before KIC runs.
 
  ## [3.4.6]